requirements.txt
hugchat
python-dotenv
gnews
newspaper4k